[Add] WebSocket chat 설정 파일 추가, studyrecord 로직 구현

+++ src/main/java/com/educoon/domain/studyRecord/StudyRecord.java
@@ -37,7 +37,7 @@ public class StudyRecord {
     private LocalDateTime endTime;
 
     @Column(nullable = false)
-    private Integer duration;
+    private Long duration;
 
 
 }
